chore: rename file (snackbar => toast)
